requirements.txt
streamlit==1.35.0
mysql-connector-python==8.3.0
pandas==2.2.2
plotly==5.21.0
python-dotenv==1.0.1
bcrypt==4.1.3
openai==1.30.1
google-generativeai==0.5.4
